+++ .idea/modules.xml
@@ -2,10 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/-EpochTimeConverter.iml" filepath="$PROJECT_DIR$/-EpochTimeConverter.iml" />
       <module fileurl="file://$PROJECT_DIR$/EpochTimeConverter.iml" filepath="$PROJECT_DIR$/EpochTimeConverter.iml" />
       <module fileurl="file://D:\AndroidApp\EpochTimeConverter\.idea\EpochTimeConverter.iml" filepath="D:\AndroidApp\EpochTimeConverter\.idea\EpochTimeConverter.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/EpochTimeConverter-app.iml" filepath="$PROJECT_DIR$/app/EpochTimeConverter-app.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
