+++ app/src/main/java/com/test/maxlukin/epochtimeconverter/Activity/MainActivity.java
@@ -1,8 +1,9 @@
-package com.test.maxlukin.epochtimeconverter;
+package com.test.maxlukin.epochtimeconverter.Activity;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import com.test.maxlukin.epochtimeconverter.R;
 
 public class MainActivity extends AppCompatActivity {
