+++ app/src/main/java/com/test/maxlukin/epochtimeconverter/Services/TimePickerFragment.java
@@ -5,7 +5,6 @@ import android.app.DialogFragment;
 import android.app.TimePickerDialog;
 import android.os.Bundle;
 import android.text.format.DateFormat;
-import android.view.View;
 import android.widget.TimePicker;
 import java.util.Calendar;
 
@@ -13,7 +12,7 @@ import java.util.Calendar;
  * Created by Zadrotisha on 28.02.2018.
  */
 
-public class TimePickerFragment extends DialogFragment
+public static class TimePickerFragment extends DialogFragment
         implements TimePickerDialog.OnTimeSetListener {
 
     @Override
@@ -27,6 +26,6 @@ public class TimePickerFragment extends DialogFragment
     }
 
     public void onTimeSet(TimePicker view, int hourOfDay, int minute) {
-
+        // Do something with the time chosen by the user
     }
 }
